helper app paths settings changed
Locally installed versions of MEME suite and HMMER package programs are now searched for first. Only if these are not found, versions included with SigmoID are used (this option may be removed as included versions don't always work).  Press 'Restore Default Paths' in the settings window to switch to these paths.